refactor(dashboard): split NxReport into Astro shell + chart-only island
Move 95% of the report panel to build-time SSG and keep only the
recharts bar chart as a React island. Same approach as Million's
"static HTML extraction" — the parts that never re-render (stat tiles,
coverage cards, plugin pills, section headings) ship as zero-JS HTML,
the only piece that actually needs React stays an island.

- `AstroNxReport.astro` reads `public/data/nx/nx-report.json` at build
  time via `import.meta.url`, parses it once, and renders the stat
  tiles, coverage card grid, and plugin pills directly as static
  HTML + scoped CSS. The recharts chart is included via
  `<ReactNxReport entries={loc} client:visible />`, so it only
  hydrates when it scrolls into view.
- `ReactNxReport.tsx` drops the orchestrator + the parsing + the
  static layout (tiles, cards, pills). It is now just the
  `<BarChart>` with its tooltip and per-language cell coloring.
- `nxReportParse.ts` is the new shared parsing module — Astro and
  React both import the same parsers, types, and color map.
- `report.mdx` swaps `<ReactNxReport client:only="react" />` for
  `<AstroNxReport />`. `ci-dashboard.yml` emits the same.

Result: the `/dashboard/report/` JS payload drops from the full
report bundle to ~1.5 KB (just the chart). The rest of the page is
pre-rendered HTML.
