+++ apps/kbve/astro-kbve/src/content/docs/dashboard/report.mdx
@@ -9,7 +9,7 @@ editUrl: false
 ---
 
 import { Card, CardGrid, Tabs, TabItem } from '@astrojs/starlight/components';
-import ReactNxReport from '@/components/dashboard/ReactNxReport';
+import AstroNxReport from '@/components/dashboard/AstroNxReport.astro';
 
 ## NX Workspace Report
 
@@ -34,7 +34,7 @@ Last generated: **2026-05-15T09:16:03Z** — updated daily by `ci-dashboard`.
   </Card>
 </CardGrid>
 
-<ReactNxReport client:only="react" />
+<AstroNxReport />
 
 ### Raw output
